Make nodesoup a private dependency of matplot++
The header `nodesoup.hpp` is only required for `axes_object/network.cpp`. Therefore it is not part of the public API. This means we don't have to install the header at all. Unfortunately, when building matplot++ as a static library CMake complains if we don't also install the nodesoup static library along with it, so we must still install `libnodesoup.a` in those cases.