Update axes_type.cpp
The plot can inherit properties of a heat map and become reversed in Y-axis and have the minor_grid enabled as proposed in the issue #110 . Adding these 2 lines to each plot function solves the issue, but it is a very bad fix in my opinion. Ideally, every class should have a feature where its default settings are reset.