fix(util): make `linspace(a,b,1)` return `a` instead of `NaN`
This underlying issue caused polarhistograms with more than 89 bins to
not be displayed at all.

Previously passing n=1 to linspace resulted in a one element vector
containing NaN. With this change it will result in a one element
vector containing the first ("lower") bound.  This is the natural
solution given the current implementation and matches the behavior of
numpy.linspace, unfortunately however it does not match the behavior
of MATLAB linspace which returns the "upper" bound instead.