fix: don't ignore `rticks()` when the r-axis is not visible
For polar plots the (unsightly, and unhelpful) r-axis is hidden by default, instead of the grid being enabled by default.

This PR handles rticks(), rtickformat(), rticklabels() and rtickangle() to generate the corresponding gnuplot commands even when the r-axis is hidden.

Additionally, it handles `rticks({})` and `rticklabels({})` properly by essentially setting `rtickformat("")`. This ensures that labels are hidden without having to disable the grid.